Refactor FotoBlurMode: Move logic to a separate file for improved organization and clarity
